Formatting (#18)
# Description

This adds the [Detekt](https://github.com/arturbosch/detekt) code analysis tool. This analysis is done solely before compilation, and will not need to be included in the production JAR.

This PR adds a tool that will conduct static analysis of the codebase to make sure the formatting and complexity of our code is up to a standard. I have gone mostly with the default configuration provided by Detekt, with a couple changes to make it align more with our project. I am open to any more suggestions on certain configuration options.

With this PR I have also included an item to the Pull Request checklist to confirm that the author of said PR has run the gradle `check` task. This task runs all of our tests and the analysis tool to confirm there are no new issues introduced. Soon, I would like to have Travis CI setup for this repository that will not only confirm that the project builds, but also that this `check` task successfully runs, but this is the first step.

## Type of change

Please delete options that are not relevant.

- [x] Bug fix (non-breaking change which fixes an issue)